moving s3 methods to new s3 singleton
